pyproject.toml
[project]
name = "templates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cheetah3>=3.2.6.post1",
    "jinja2>=3.1.4",
    "mako>=1.3.6",
]
